ensure subprocesses write to separate files and remove excess logging to prevent hangs (#235)
The python3 multiprocessing pool map function has a number of performance drawbacks that make it ill suited for this particular type of production code.  It's better to output subprocess results to files and concatenate later, rather than serialize and pipe through map results.  The performance impact of this would be felt on samples with tens of thousands of contigs.  We recently ran a sample with 169,000+ contigs.  The JSON output format for the pileup data is also a performance liability (costs a lot of RAM to load the entire JSON structure before its processing can begin) but where it is consumed downstream this pattern is deeper rooted, so leaving it in JSON for now.

The specific motivator for this change was a deadlock observed on accessing logging concurrently from subprocesses.

Also including here a change to the run_in_subprocess logging to ensure that any logging occurs in the parent process, not in the subprocess;  again this should minimize the potential for deadlock.