Prepare to migrate JSON files out of idseq-dag (#80)
With the miniwdl sfn-wdl plugin enhanced to create similar status JSON files alongside, plumb some information for that to use (dynamically-generated step descriptions). Once the webapp is consuming the new versions, we can disable/delete the corresponding idseq-dag routines. (This diff deletes a few bits of it that are already dead code.)